_for_deleting/requirements.txt
altair==5.5.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.19.0
nltk==3.9.1
numpy==2.2.1
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.29.2
pyarrow==18.1.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.1
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
tweepy @ git+https://github.com/tweepy/tweepy.git@91a41c6e1c955d278c370d51d5cf43b05f7cd979
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
openai>=1.0.0
python-dotenv==1.0.0
streamlit==1.29.0
ntscraper==0.3.0
sqlite3